infrastructure/django/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.1.20
django-cors-headers==4.4.0
influxdb-client==1.49.0
PyYAML==6.0.2
